better coverage for TTL testing
